More documentation, and removed unnecessary binaries
